1- Rename namespace of ApiClient from "commercetools.Sdk.Domain.ApiClient" to ""commercetools.Sdk.Domain.ApiClients" 2- Rename namespace of customObject Models from "commercetools.Sdk.Domain.CustomObject" to "commercetools.Sdk.Domain.CustomObjects" 3- Rename namespace of InventoryEntry Models from "commercetools.Sdk.Domain" to "commercetools.Sdk.Domain.InventoryEntries" 4- Rename namespace of InventoryEntry UpdateActions from "commercetools.Sdk.Domain.InventoryEntries" to "commercetools.Sdk.Domain.InventoryEntries.UpdateActions"
