+++ commercetools.Sdk/IntegrationTests/commercetools.Sdk.IntegrationTests/CustomObjects/CustomObjectsIntegrationTests.cs
@@ -2,7 +2,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using commercetools.Sdk.Client;
 using commercetools.Sdk.Domain;
-using commercetools.Sdk.Domain.CustomObject;
+using commercetools.Sdk.Domain.CustomObjects;
 using commercetools.Sdk.HttpApi.Domain.Exceptions;
 using commercetools.Sdk.HttpApi.Extensions;
 using Xunit;
