+++ commercetools.Sdk/commercetools.Sdk.Client/DeleteByContainerAndKeyCommand.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using commercetools.Sdk.Domain;
 using commercetools.Sdk.Domain.Carts;
-using commercetools.Sdk.Domain.CustomObject;
 using commercetools.Sdk.Domain.Orders;
 using commercetools.Sdk.Domain.Query;
