--- commercetools.Sdk/commercetools.Sdk.Domain/ApiClient/ApiClient.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace commercetools.Sdk.Domain.ApiClient
-{
-    [Endpoint("api-clients")]
-    public class ApiClient
-    {
-    }
-}
