+++ commercetools.Sdk/commercetools.Sdk.Domain/ApiClients/ApiClient.cs
@@ -1,0 +1,7 @@
+﻿namespace commercetools.Sdk.Domain.ApiClients
+{
+    [Endpoint("api-clients")]
+    public class ApiClient
+    {
+    }
+}
