--- commercetools.Sdk/commercetools.Sdk.Domain/CustomObject/CustomObject.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace commercetools.Sdk.Domain.CustomObject
-{
-    public class CustomObject : CustomObjectBase
-    {
-        public object Value { get; set; }
-    }
-}
