+++ commercetools.Sdk/commercetools.Sdk.Domain/CustomObjects/CustomObject.cs
@@ -1,0 +1,7 @@
+﻿namespace commercetools.Sdk.Domain.CustomObjects
+{
+    public class CustomObject : CustomObjectBase
+    {
+        public object Value { get; set; }
+    }
+}
