+++ commercetools.Sdk/commercetools.Sdk.Domain/InventoryEntries/UpdateActions/AddQuantityUpdateAction.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace commercetools.Sdk.Domain.InventoryEntries
+namespace commercetools.Sdk.Domain.InventoryEntries.UpdateActions
 {
     public class AddQuantityUpdateAction : UpdateAction<InventoryEntry>
     {
