+++ commercetools.Sdk/commercetools.Sdk.Domain/InventoryEntries/UpdateActions/ChangeQuantityUpdateAction.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace commercetools.Sdk.Domain.InventoryEntries
+namespace commercetools.Sdk.Domain.InventoryEntries.UpdateActions
 {
     public class ChangeQuantityUpdateAction : UpdateAction<InventoryEntry>
     {
