+++ commercetools.Sdk/commercetools.Sdk.Domain/InventoryEntries/UpdateActions/RemoveQuantityUpdateAction.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace commercetools.Sdk.Domain.InventoryEntries
+namespace commercetools.Sdk.Domain.InventoryEntries.UpdateActions
 {
     public class RemoveQuantityUpdateAction : UpdateAction<InventoryEntry>
     {
