+++ commercetools.Sdk/commercetools.Sdk.Domain/InventoryEntries/UpdateActions/SetExpectedDeliveryUpdateAction.cs
@@ -1,7 +1,6 @@
 ﻿using System;
-using System.ComponentModel.DataAnnotations;
 
-namespace commercetools.Sdk.Domain.InventoryEntries
+namespace commercetools.Sdk.Domain.InventoryEntries.UpdateActions
 {
     public class SetExpectedDeliveryUpdateAction : UpdateAction<InventoryEntry>
     {
