+++ commercetools.Sdk/commercetools.Sdk.Domain/Payments/UpdateActions/AddInterfaceInteractionUpdateAction.cs
@@ -1,5 +1,4 @@
 ﻿using System.ComponentModel.DataAnnotations;
-using commercetools.Sdk.Domain.CustomObject;
 
 namespace commercetools.Sdk.Domain.Payments.UpdateActions
 {
