+++ commercetools.Sdk/commercetools.Sdk.HttpApi/RequestBuilders/DeleteByContainerAndKeyMessageBuilder.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Net.Http;
 using commercetools.Sdk.Client;
-using commercetools.Sdk.Domain.CustomObject;
 using commercetools.Sdk.HttpApi.AdditionalParameters;
 using commercetools.Sdk.Serialization;
 using Microsoft.AspNetCore.WebUtilities;
